wilc3000-external-module: multiple fixups for cleaner builds
Previously, FILES was not specified in the recipe. This was normally
not needed but it would appear meta-openembedded adds for tighter
QA restrictions on this.

The potential for usrmerge also complicated the build and files.
This was addressed with some magic for INSTALL_MOD_PATH

EXTRA_OEMAKE variables were speficied for the variables such as
M, CONFIG_WILC_SPI, etc. normally needed at build time.

Last, the B directory was updated to the actual builddir rather
than just the staging dir. This allows for complete removal of
custom do_compile and do_install commands, letting the module.bbclass
do all of the heavy lifting

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>